refactor: Simplified code
