+++ src/components/button/button.tsx
@@ -1,17 +1,23 @@
 import { ReactNode, useState } from "react";
-import { constantsButton } from ".";
 import { constants } from "../../typography";
 import { colors } from "../../typography/colors";
 
+export const constantsButton = {
+  defaultCompactHeight: "40px",
+  defaultHeight: "48px",
+  boxShadow: "0px 4px 4px rgba(0, 0, 0, 0.25)",
+};
+
 const stylesButton = {
-  minWidth: constantsButton.minWidth,
+  padding: "16px",
+  minWidth: "143px",
   fontFamily: constants.font.family.primary,
-  fontWeight: constantsButton.weight,
+  fontWeight: constants.font.weight.bold,
   borderRadius: constants.font.radius,
-  cursor: constantsButton.cursor,
-  display: constantsButton.display,
-  justifyContent: constantsButton.justifyContent,
-  alignItems: constantsButton.alignItems,
+  cursor: "pointer",
+  display: "flex",
+  justifyContent: "center",
+  alignItems: "center",
 };
 
 interface ButtonProps {
@@ -41,24 +47,33 @@ export function ButtonPrimary({
     setHoveractive(false);
   };
 
+  let backgroundColor, boxShadow, height, border, color;
+
+  if (variant === "primary") {
+    backgroundColor = disabled ? colors.NeutralLight : colors.PrimaryXDark;
+  } else if (variant === "CTA") {
+    backgroundColor = colors.CTA;
+  } else if (variant === "secondary") {
+    backgroundColor = colors.Neutral;
+    border = `2px solid ${colors.Accessory1}`;
+    color = colors.Accessory1;
+  } else {
+    backgroundColor = colors.PrimaryXDark;
+  }
+  boxShadow = hoveractive ? constantsButton.boxShadow : "none";
+  height = compact
+    ? constantsButton.defaultCompactHeight
+    : constantsButton.defaultHeight;
+  border = variant === "secondary" ? `2px solid ${colors.Accessory1}` : "none";
+  color = variant === "secondary" ? colors.Accessory1 : colors.Neutral;
+
   const styles = {
     ...stylesButton,
-    backgroundColor:
-      variant === "primary"
-        ? disabled
-          ? colors.NeutralLight
-          : colors.PrimaryXDark
-        : variant === "CTA"
-        ? colors.CTA
-        : variant === "secondary"
-        ? colors.Neutral
-        : colors.PrimaryXDark,
-    boxShadow: hoveractive ? constantsButton.boxShadow : "none",
-    height: compact
-      ? constantsButton.defaultCompactHeight
-      : constantsButton.defaultHeight,
-    border: variant === "secondary" ? `2px solid ${colors.Accessory1}` : "none",
-    color: variant === "secondary" ? colors.Accessory1 : colors.Neutral,
+    backgroundColor,
+    boxShadow,
+    height,
+    border,
+    color,
   };
 
   return (
