--- src/components/button/index.ts
@@ -1,14 +1,0 @@
-import { constants } from "../../typography";
-
-export const constantsButton = {
-  minWidth: "143px",
-  defaultCompactHeight: "40px",
-  defaultHeight: "48px",
-  weight: constants.font.weight.bold,
-  boxShadow: "0px 4px 4px rgba(0, 0, 0, 0.25)",
-  cursor: "pointer",
-  padding: "16px",
-  display: "flex",
-  justifyContent: "center",
-  alignItems: "center",
-};
